Circpad event log negotiation for relays
This allows relays to log circpad cell traces for researcher controlled
circuits (and only those circuits).

It works by sending a single padding negotiate cell from the client, that
tells the relay to log, and includes the client circid. The relay does not
reply.

This single cell is followed by a circpad_negotiate_logging event on both
sides. The post-processor should remove the nonpadding_sent following the
circpad_negotiate_logging event on the client side, and remove the only the
circpad_negotiate_logging event on the relay. The circpad_negotiate_logging
event should be the first log message for a circuit on a relay.

This logging negotiation only happens on the live network. Log negotiation is
disabled in the simulator via the TOR_UNIT_TESTS ifndef. This means the
simulator will never send these log cells.