Provide batching support for circpad sim
With this, you can specify directories of traces instead of trace files.

If you specify files, it detects this and works as before, so this branch
should still be backwards compatible with the old python.

If you specify directories, it checks that the same number of files are in
each directoty, and then takes the list of files and loads a trace from the
client directory, and a trace from the relay directory.

Before each trace, a logline of BATCH_TRACE=filename is emitted.
The circuitid is also incremented after each run.

Because this only runs on one CPU core, for large amounts of traces, you may
want to separate your directories into a set of pairs of directories, with one
directory pair for each CPU core. Then, you can launch this in parallel for
all pairs, to utilize all CPU cores. Perhaps with some python, this
dataset splitting and launch parallelization can be done automatically.

Timestamps appear to be working properly, but please verify this behavior
closely, if time is used as a feature. Also sanity check that this produces
the same output as the single-shot version, as I have not done a detailed
comparison, but it looks right to me. I also did not test the event cutoff
behavior.